+++ build_runner_core/lib/src/asset_graph/serializers.g.dart
@@ -8,13 +8,24 @@ part of 'serializers.dart';
 
 Serializers _$serializers =
     (new Serializers().toBuilder()
+          ..add(AssetDeps.serializer)
           ..add(AssetNode.serializer)
+          ..add(ExpiringValue.serializer)
           ..add(GeneratedNodeConfiguration.serializer)
           ..add(GeneratedNodeState.serializer)
           ..add(GlobNodeConfiguration.serializer)
           ..add(GlobNodeState.serializer)
           ..add(NodeType.serializer)
+          ..add(PhasedAssetDeps.serializer)
+          ..add(PhasedValue.serializer)
           ..add(PostProcessBuildStepId.serializer)
+          ..addBuilderFactory(
+            const FullType(BuiltMap, const [
+              const FullType(AssetId),
+              const FullType(PhasedValue, const [const FullType(AssetDeps)]),
+            ]),
+            () => new MapBuilder<AssetId, PhasedValue<AssetDeps>>(),
+          )
           ..addBuilderFactory(
             const FullType(BuiltSet, const [const FullType(AssetId)]),
             () => new SetBuilder<AssetId>(),
@@ -26,6 +37,14 @@ Serializers _$serializers =
           ..addBuilderFactory(
             const FullType(BuiltList, const [const FullType(AssetId)]),
             () => new ListBuilder<AssetId>(),
+          )
+          ..addBuilderFactory(
+            const FullType(BuiltSet, const [const FullType(AssetId)]),
+            () => new SetBuilder<AssetId>(),
+          )
+          ..addBuilderFactory(
+            const FullType(BuiltSet, const [const FullType(AssetId)]),
+            () => new SetBuilder<AssetId>(),
           )
           ..addBuilderFactory(
             const FullType(BuiltSet, const [const FullType(AssetId)]),